refactor: breakup event listeners
