+++ src/main/kotlin/me/tech/mcchestui/listeners/item/GUIItemPlaceListener.kt
@@ -1,0 +1,60 @@
+package me.tech.mcchestui.listeners.item
+
+import me.tech.mcchestui.GUI
+import me.tech.mcchestui.listeners.GUIEventListener
+import org.bukkit.Material
+import org.bukkit.entity.Player
+import org.bukkit.event.EventHandler
+import org.bukkit.event.inventory.InventoryAction
+import org.bukkit.event.inventory.InventoryClickEvent
+
+internal class GUIItemPlaceListener(gui: GUI) : GUIEventListener(gui) {
+    @EventHandler
+    internal fun InventoryClickEvent.itemPlace() {
+        // ensure top inventory is ui inventory.
+        if(!gui.isBukkitInventory(inventory)) {
+            return
+        }
+
+        if(
+            action == InventoryAction.MOVE_TO_OTHER_INVENTORY
+            && isShiftClick
+            && !gui.isBukkitInventory(clickedInventory) // make sure its incoming.
+        ) {
+            if(!gui.allowItemPlacement) {
+                isCancelled = true
+                return
+            }
+        } else if(
+            action in PLACE_ACTIONS
+            && gui.isBukkitInventory(clickedInventory)
+        ) {
+            if(!gui.allowItemPlacement) {
+                isCancelled = true
+                return
+            }
+        } else {
+            return
+        }
+
+        val guiSlot = gui.slots.getOrNull(slot)
+        if(guiSlot != null) {
+            if(!guiSlot.allowPickup) {
+                isCancelled = true
+                return
+            }
+        }
+
+        val itemStack = cursor
+            ?: return
+        if(itemStack.type == Material.AIR) {
+            return
+        }
+
+        gui.onPlaceItem?.let { uiEvent ->
+            uiEvent(this, whoClicked as Player, itemStack, slot).let { outcome ->
+                isCancelled = outcome
+            }
+        }
+    }
+}
